Fix failure of scons build for android.
It was shown error message about std::to_string()
when building for android.

So we change some API' parameter for android.

Reviewed-on: https://gerrit.iotivity.org/gerrit/577
Tested-by: Uze Choi <uzchoi@samsung.com>

+++ service/things-manager/sdk/inc/ThingsManager.h
@@ -413,7 +413,7 @@ namespace OIC
          * NOTE: OCStackResult is defined in ocstack.h.
          */
         OCStackResult executeActionSet(std::shared_ptr< OCResource > resource,
-                std::string actionsetName, long int delay, PostCallback cb);
+                std::string actionsetName, unsigned int delay, PostCallback cb);
 
         /**
          * API for canceling an existing action set.
